deps: move postman-request to direct dep for smoother upgrades.
When upgrading from 7 to 8, npm and classic Yarn
would not automatically install postman-request because
it's a peer-dep and not a direct dep.

By making postman-request a direct dep, upgrades will
go smoother.

To minimize the need to upgrade this module due to a
vuln that's found later in postman-request, we intentionally
use the `^` notation to accept any newer patch version,
so users would be able to update their lock files without
requiring a new release of this module.